+++ crispy_forms/tests/test_layout.py
@@ -4,7 +4,10 @@ from __future__ import unicode_literals
 from django import forms
 from django.core.urlresolvers import reverse
 from django.forms.models import formset_factory, modelformset_factory
-from django.middleware.csrf import _get_new_csrf_key
+try:
+    from django.middleware.csrf import _get_new_csrf_key
+except ImportError:
+    from django.middleware.csrf import _get_new_csrf_string as _get_new_csrf_key
 from django.shortcuts import render_to_response
 from django.template import (
     Context, RequestContext
@@ -300,8 +303,6 @@ def test_formset_layout(settings):
     assert html.count('Note for first form only') == 1
     if settings.CRISPY_TEMPLATE_PACK == 'uni_form':
         assert html.count('formRow') == 3
-    elif settings.CRISPY_TEMPLATE_PACK == 'bootstrap4':
-        assert html.count('row') == 21
     else:
         assert html.count('row') == 3
 
@@ -537,9 +538,7 @@ def test_keepcontext_context_manager(settings):
         InlineCheckboxes('alphacheckboxes'),
         'numeric_multiple_checkboxes'
     )
-    request_factory = RequestFactory()
-    request = request_factory.get('/')
-    context = RequestContext(request, {'form': form})
+    context = {'form': form}
 
     response = render_to_response('crispy_render_template.html', context)